Split apart test files
